Moved inex.js to demofile-parsing folder
ALSO - added Testing boolean that can be turned on and off
